Polarity Table of Translated AFINN Based Lexicon

The polarity table of AFINN-111 based lexicon translated by Roman Yudin (Node.js package SentiMental) and augmented by Alexey Romanov from Text Machine Lab team (Python library Sentimental). Table contains 7268 words with non-neutral sentiment scores.

data(hash_sentiment_afinn_ru)

Format

A data table with 7268 rows and 2 variables:

token

the textual token (word or phrase)

score

the sentiment score: from −5 (negative) to +5 (positive) with fractional step

Source

https://github.com/text-machine-lab/sentimental/blob/master/sentimental/word_list/russian.csv

Details

The augmentation of Text Machine Lab team is the words having fractional scores while original AFINN lexicon assigns to words only integer scores. The duplicated words was removed by rulexicon maintainer.

License

The original AFINN-111 lexicon published under "Open Database License (ODbL) v1.0" license or a similar copyleft license.

References

AFINN lexicon: http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010

Node.js package SentiMental: https://github.com/Wobot/Sentimental

Python library Sentimental: https://github.com/text-machine-lab/sentimental

Text Machine Lab of University of Massachusetts Lowell: http://text-machine.cs.uml.edu/lab