Polarity Table of Chen and Skiena (2014) Lexicon Multilingualsentiment

The polarity table of Chen and Skiena's (2014) lexicon Multilingualsentiment. Table contains 2876 words with non-neutral sentiment scores.

data(hash_sentiment_chen_skiena)

Format

A data table with 2876 rows and 2 variables:

token

the textual token (word or phrase)

score

the sentiment score: −1 for negative, 1 for positive

Source

http://bit.ly/1iwOMZb - negative words list, http://bit.ly/1rAqgfb - positive words list

Details

The lexicon was builded on the basis of knowledge graph that connect words in different languages (based on Wiktionary, Google Translate, transliteration links and WordNet). The authors provide two data files: one for negative and one for positive words. This polarity table is a joined list after deletion of some artifacts (36 words in English for which however the translations are in data files).

License

The lexicon is distributed under the GNU General Public License (v3 or later). The authors note that this is the full GPL, which allows many free uses, but does not allow its incorporation into any type of distributed proprietary software, even in part or in translation. For commercial applications please contact the lexicon creators.

References

Chen, Y., Skiena, S., 2014. Building Sentiment Lexicons for All Major Languages. In Proceedings of 52nd Annual Meeting of the Association for Computational Linguistics, pp. 383-389.

Multilingualsentiment project web-page: https://sites.google.com/site/datascienceslab/projects/multilingualsentiment