Polarity Table of Sentiment-Thematic Dictionary of Economic Terms - EcSentiThemeLex

The polarity table of sentiment-thematic dictionary of economic, financial and legal terms - EcSentiThemeLex. Table contains 15235 words forms, and 9841 has non-neutral sentiment scores.

data(hash_sentiment_ecsentithemelex)

Format

A data table with 15235 rows and 2 variables:

token

a token (word)

score

a sentiment score: from −2 (strongly negative) to +2 (strongly positive) with step 1

Details

The polarity table was generated from original lexicon table (see key_ecsentithemelex) based on the following rules:

tokens consisting of more than one word were discarded (as required for bag-of-words sentiment analysis)

all possible word forms were added as separete tokens

for the words containing the letter "ё" the spelling option with letter "е" were added as separete token

License

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0). For additional permissions (including the commercial use) please contact to Elena Fedorova <ecolena@mail.ru>.

References

Fedorova, E., Afanasyev, D., Demin, I., Lazarev, A., Nersesyan, R., Pyltsin, I.V. (2020). Development of a tonal-thematic dictionary EcSentiThemeLex for the analysis of economic texts in Russian. Journal of Applied Informatics, 6 (15), 58–77. DOI: https://doi.org/10.37791/2687-0649-2020-15-6-58-77.