Polarity Table of Social-Political Oriented Lexicon "LINIS Crowd SENT" 2016

The polarity table from crowdsourcing lexicon "LINIS Crowd SENT", version 2016. Table contains 6860 words or phrases, and 2454 has non-neutral sentiment scores.

data(hash_sentiment_linis_crowd_2016)

Format

A data table with 6860 rows and 2 variables:

token

the textual token (word or phrase)

score

the sentiment score: from −2 (negative) to +2 (positive) with step 1

Source

http://linis-crowd.org/static/collection_(docs_words)_2016_all_labels.zip

Details

The lexicon is aimed at detecting sentiment in blogs and social media related to social and political issues. Words sentiment scores assessed by volunteers (at least three). Source file provides raw scores (from each volunteers) and were averaged and rounded to the nearest integer by maintainer of rulexicon package.

License

According to "LINIS Crowd SENT" project web-site (http://linis-crowd.org) the dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International (CC BY-NC-SA 4.0). Additional permissions can be accessed here: http://www.linis-crowd.org/contacts/.

References

Koltsova, O.Yu., Alexeeva, S.V., Kolcov S.N., 2016. An opinion word lexicon and a training dataset for Russian sentiment analysis of social media. Computational Linguistics and Intellectual Technologies: Proceedings of the International Conference "Dialogue 2016". URL: http://www.dialog-21.ru/media/3400/koltsovaoyuetal.pdf