Polarity Table of Translated NRC Emotion Lexicon

The polarity table of filtered and translated version of Mohammad & Turney's (2010) positive/negative word list. Table contains 5179 words with non-neutral sentiment scores.

data(hash_sentiment_nrc_emolex_ru)

Format

A data table with 5179 rows and 2 variables:

token

the textual token (word or phrase)

score

the sentiment score: −1 for negative, 1 for positive

Source

http://sentiment.nrc.ca/lexicons-for-research/NRC-Emotion-Lexicon.zip

Details

The Russian translation is based on the original word translations file provided by lexicon's creators.

License

The original authors note the data is available for non-commercial use. If you are interested in commercial use of lexicon: "... send email to Saif M. Mohammad (Senior Research Officer at NRC and creator of these lexicons): saif.mohammad@nrc-cnrc.gc.ca and Pierre Charron (Client Relationship Leader at NRC): Pierre.Charron@nrc-cnrc.gc.ca. A nominal one-time licensing fee may apply."

References

Mohammad, S.M. and Turney, P.D. (2010). Emotions evoked by common words and phrases: Using Mechanical Turk to create an emotion lexicon. In Proceeding of Workshop on Computational Approaches to Analysis and Generation of Emotion in Text, pp. 26-34.

http://sentiment.nrc.ca/lexicons-for-research/