Dictionary of Economic Terms of Tochka Bank

A dataset containing the dictionary of financial, economic and legal terms based on a glossary of Tochka Branch of "Bank Otkritie Financial Corporation" (Public Joint-Stock Company), and augmented by Dmitriy O. Afanasyev. Table contains 6058 words and phrases.

Usage

data(key_economic_tochka_bank)

Format

A data frame with 6058 rows and 3 variables:

token

a token (term)

lemma

a lemma of token (term)

regex

a regular expression of the all possible word forms of the lemma

Source

https://tochka.com/info/glossary/

Details

The dictionary is a glossary of Tochka Branch of "Bank Otkritie Financial Corporation" (Public Joint-Stock Company) https://tochka.com/info/glossary/ based on Borisov's "Great economic dictionary" and augmented by maintainer of rulexicon package in the following:

fixed some issues with "е" instead "ё" and others

lemmatized form of terms was added using hash_lemmas_opencorpora (the lemma column)

the different word forms were added as universal regular expression, including the dualism of "ё - е" (the regex column)

Along with traditional terms, the dictionary includes relatively new ones that have come into use in recent years. The dictionary can be used to classify the texts as "economic" by simple terms search and counting.

License

There is no explicit information about the license politics on a web-site Tochka Branch of "Bank Otkritie Financial Corporation" (Public Joint-Stock Company) or in a book of Borisov A.B. It can be assumed that the dictionary can be published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0).

References

Glossary of Tochka Branch of "Bank Otkritie Financial Corporation" (Public Joint-Stock Company): https://tochka.com/info/glossary/

Borisov, A.B. (2006). Great economic dictionary. Knizhiy Mir, Moscow. 543 p.