Dictionary of Innovation Activities Disclosure of Garechana et al. and Libaers et al. (English)

A dataset containing the dictionary (taxonomy) for innovation activities disclosure estimation in companies’ annual reports based on lexicons from Garechana et al. (2017) and Libaers et al. (2016). Table contains 41 words and phrases.

data(key_innov_activ_en)

Format

A data frame with 41 rows and 3 variables:

main_token

main or grouping token (word or phrase)

token

token (word or phrase), this is equal to or synonym of main_token

regex

regular expression for all word forms of the token

Details

The dictionary enables to estimate the level of innovation disclosure in companies’ annual reports. Lexicons from Garechana et al. 2017 and Libaers et al. 2016) were used as a basis.

Languages

Russian: key_innov_activ

License

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0).

References

Garechana, G., Río-Belver, R., Bildosola, I., Rodríguez-Salvador, M. (2017). Effects of innovation management system standardization on firms: evidence from text mining annual reports. Scientometrics, 111(3), 1987–1999. DOI: https://doi.org/10.1007/s11192-017-2345-7.

Libaers, D., Hicks, D., Portery, A.L. (2016). A taxonomy of small firm technology commercialization. Industrial and Corporate Change, 25(3), 371–405. DOI: https://doi.org/10.1093/icc/dtq039.