Dictionary of innovation activities disclosure by OECD types of innovations

A dataset containing the dictionary (taxonomy) for innovation activities disclosure estimation in companies’ annual reports according to the OECD Oslo Manual's innovation types.The dictionary builded by Fedorova et al. Table contains 65 words and phrases.

data(key_innov_activ_oecd)

Format

A data frame with 65 rows and 3 variables:

type

an innovation type according to the OECD Oslo Manual

token

a token (word or phrase)

regex

regular expression for all word forms of the token (including "е-ё" dualism)

Details

The dictionary enables to estimate the level of disclosure of certain types of innovations in companies’ annual reports. Each word was evaluated in terms of its relevance to a certain type of innovation according to the OECD Oslo Manual:

Продуктовые инновации; Процессные инновации; Маркетинговые инновации; Организационные инновации .

Languages

English: key_innov_activ_oecd_en

License

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0). For additional permissions (including the commercial use) please contact to Elena Fedorova <ecolena@mail.ru>.

References

OECD Oslo Manual: https://www.oecd.org/science/oslo-manual-2018-9789264304604-en.htm