Dictionary of Non-financial Information Disclosure of Fedorova, Khrustova and Demin (GRI based)

A dataset containing the dictionary (taxonomy) for non-financial information disclosure estimation based on Global Reporting Initiative Standards glossary (GRI, https://www.globalreporting.org/standards/), augmented by Fedorova, Khrustova and Demin. Table contains 340 words and phrases.

data(key_nonfinance_report_standard)

Format

A data frame with 340 rows and 5 variables:

order

number of report's section (area) order

section

name of report's section (area)

main_token

main or grouping token (word or phrase)

token

token (word of phrase), this is equal to or synonym of main_token

regex

regular expression for all word forms of the token

Source

https://www.globalreporting.org/standards/gri-standards-download-center/

Details

The dictionary is based on Global Reporting Initiative Standards glossary (https://www.globalreporting.org/standards/) augmented by Fedorova, Khrustova and Demin in the following:

terms of general nature were deleted

terms were extended and aligned with the terminology of Russian companies' reports

glossary was supplemented by additional topics and related terms: sponsorship, social responsibility, innovativeness, technology

synonyms and different word forms were added as universal regular expression

The dictionary includes the report's sections (areas) as following:

Основные элементы; Экономическая результативность; Экология; Занятость; Социальная ответственность; Благотворительная и спонсорская деятельность; Инновационность и научно-техническое развитие .

The taxonomy can be used to verify compliance of Russian public company reports with GRI Standards, non-financial information disclosure estimation, as well as for thematic analysis of company reports. See References for details of usage.

License

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0). For additional permissions (including the commercial use) please contact to Elena Fedorova <ecolena@mail.ru>.

References

Fedorova, E.A., Khrustova, L.E., Demin I.S. (2020). The impact of the completeness of non-financial information disclosure by Russian companies on investment attractiveness. Russian Management Journal 18(1). Forthcoming.