Profane Words List of bally_36 LiveJournal user

A dataset containing a character vector of profane words from LiveJournal user bally_36 augmented by maintainer of rulexicon package using additional list from www.spy-soft.net user IVO.

data(profanity_bally)

Format

A character vector with 825 elements.

Source

Complex source, see References

License

There is no license informations provided by LiveJournal user bally_36 or www.spy-soft.net user IVO explicitly. But content was published for free reusage.

References

LiveJournal bally_36 users profane words list: https://bally-36.livejournal.com/66906.html

www.spy-soft.net IVO users profane words list: http://www.spy-soft.net/slovar-dlya-bruta-russkij-mat