Dictionary of Public Company Information Disclosure of Fedorova, Grishenko and Demin (Regulation No. 454-P)

A dataset containing the dictionary (taxonomy) of Fedorova, Grishchenko and Demin for the public company information disclosure estimation based on Regulation No. 454-P of the Central Bank of Russian Federation (https://www.cbr.ru/Queries/XsltBlock/File/86277?fileid=-1&scope=1614-1615). Table contains 189 words and phrases.

data(key_company_rus_report_standard)

Format

A data frame with 189 rows and 5 variables:

order

number of report's section (area) order

section

name of report's section (area)

main_token

main or grouping token (word or phrase)

token

token (word or phrase), this is equal to or synonym of main_token

regex

regular expression for all word forms of the token

Details

The dictionary is based on Regulation No. 454-P of the Central Bank of Russian Federation (https://www.cbr.ru/Queries/XsltBlock/File/86277?fileid=-1&scope=1614-1615) and builded by Fedorova, Grishenko and Demin:

the main terms of Russian companies' reports were manually collected into dictionary

synonyms and different word forms were added as universal regular expression.

The dictionary includes the report's sections (areas) as following:

Характеристика деятельности компании; Ценные бумаги и дивидендная политика; Сведения о сделках; Органы управления; Корпоративное управление (принципы); Органы контроля; Управление рисками; Прочее, обязательное раскрытие .

The taxonomy can be used to verify compliance of Russian public company reports with standards of the Central Bank of Russian Federation, information disclosure estimation, as well as for thematic analysis of company reports. See References for details of usage.

License

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0). For additional permissions (including the commercial use) please contact to Elena Fedorova <ecolena@mail.ru>.

References

Fedorova, E.A., Grishenko, U.I., Demin I.S. (2020). The impact of the level of disclosure of mandatory non-financial information in the annual reports of public companies on the investment attractiveness of the issuer. Forthcoming.